｢subpc-x13｣での変更を反映

+++ agc029/a/main_2.cpp
@@ -12,7 +12,7 @@ using ull = unsigned long long;
 const int INF = 1e9;
 const ll LLINF = 1e18;
 
-int dp[200010];
+ll dp[200010];
 
 int main(){
     string s; cin >> s;
